backend/requirements.txt
fastapi[standard]==0.115.12
pymongo==4.12.1
uvicorn==0.34.2
python-dotenv==1.1.1
requests==2.32.4